restype.scala
